Move code to align.multiple package and rename classes #278
