+++ biojava-structure-gui/src/main/java/org/biojava/nbio/structure/align/gui/aligpanel/MultStatusDisplay.java
@@ -29,10 +29,10 @@ import java.util.List;
 import javax.swing.JTextField;
 
 import org.biojava.nbio.structure.Atom;
-import org.biojava.nbio.structure.align.cemc.MultipleAlignmentTools;
 import org.biojava.nbio.structure.align.gui.DisplayAFP;
 import org.biojava.nbio.structure.align.gui.jmol.JmolTools;
-import org.biojava.nbio.structure.align.model.MultipleAlignment;
+import org.biojava.nbio.structure.align.multiple.MultipleAlignment;
+import org.biojava.nbio.structure.align.multiple.MultipleAlignmentTools;
 import org.biojava.nbio.structure.gui.events.AlignmentPositionListener;
 import org.biojava.nbio.structure.gui.util.AlignedPosition;
